backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.12.0
httpx==0.26.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pandas==2.2.0
openpyxl==3.1.2
matplotlib==3.8.2
chromadb==0.4.22
langchain==0.1.5
langchain-openai==0.0.5
langchain-community>=0.0.17,<0.1
PyPDF2==3.0.1
python-docx==1.1.0
websockets==12.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
